Fix a bug where gevent.subprocess.Popen was making stdin blocking again
This was subtle, and was a bitch to find.
So, when running on a standard terminal, fds 0 and 2 point to the same
file entry. And in Popen, in the child before exec, the new stdin/out/err
have their nonblock flags cleared. And stderr was being passed through.
Hence, it was clearing stderr's nonblock flag, which cleared stdin's.

The workaround here is to redirect stderr to /dev/null.